feat(redwoodjs): migrate Audiences examples to Segments API
