+++ redwoodjs-resend-examples/javascript/web/src/layouts/MainLayout/MainLayout.jsx
@@ -46,8 +46,8 @@ const MainLayout = ({ children }) => {
         <Link to="/domains" style={navLinkStyle}>
           Domains
         </Link>
-        <Link to="/audiences" style={navLinkStyle}>
-          Audiences
+        <Link to="/segments" style={navLinkStyle}>
+          Segments
         </Link>
       </nav>
       <main>{children}</main>
