+++ redwoodjs-resend-examples/typescript/api/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "resend": "^4.0.0",
+    "resend": "^6.24.0",
     "svix": "^1.0.0"
   }
 }
