+++ redwoodjs-resend-examples/typescript/api/src/functions/doubleOptinWebhook.ts
@@ -26,11 +26,11 @@ export const handler = async (event: APIGatewayEvent, _context: Context) => {
     };
   }
 
-  const audienceId = process.env.RESEND_AUDIENCE_ID;
-  if (!audienceId) {
+  const segmentId = process.env.RESEND_SEGMENT_ID;
+  if (!segmentId) {
     return {
       statusCode: 500,
-      body: JSON.stringify({ error: "RESEND_AUDIENCE_ID not configured" }),
+      body: JSON.stringify({ error: "RESEND_SEGMENT_ID not configured" }),
     };
   }
 
@@ -64,7 +64,7 @@ export const handler = async (event: APIGatewayEvent, _context: Context) => {
     }
 
     // Find the contact by email
-    const { data: contacts } = await resend.contacts.list({ audienceId });
+    const { data: contacts } = await resend.contacts.list({ segmentId });
     const contact = contacts?.data.find((c) => c.email === recipientEmail);
 
     if (!contact) {
@@ -76,7 +76,7 @@ export const handler = async (event: APIGatewayEvent, _context: Context) => {
 
     // Update contact: confirm subscription
     await resend.contacts.update({
-      audienceId,
+      segmentId,
       id: contact.id,
       unsubscribed: false,
     });
