+++ redwoodjs-resend-examples/typescript/web/src/pages/SegmentsPage/SegmentsPage.tsx
@@ -3,24 +3,20 @@ import { MetaTags } from "@redwoodjs/web";
 import { PageHeader } from "src/components/PageHeader";
 import { ResultDisplay } from "src/components/ResultDisplay";
 
-const buttonStyle = { padding: "10px 20px", backgroundColor: "#18181b", color: "#fff", border: "none", borderRadius: "6px", fontSize: "14px", fontWeight: 500, cursor: "pointer" };
-const codeBlockStyle = { marginTop: "32px", padding: "20px", backgroundColor: "#18181b", borderRadius: "8px", color: "#e4e4e7" };
-const preStyle = { margin: 0, fontSize: "13px", whiteSpace: "pre-wrap", wordBreak: "break-all" };
-
-const AudiencesPage = () => {
+const SegmentsPage = () => {
   const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState(null);
+  const [result, setResult] = useState<Record<string, unknown> | null>(null);
 
   const fetchContacts = async () => {
     setLoading(true);
     setResult(null);
 
     try {
-      const res = await fetch("/.redwood/functions/audiencesContacts");
+      const res = await fetch("/.redwood/functions/segmentsContacts");
       const data = await res.json();
       setResult(data);
     } catch (err) {
-      setResult({ error: err.message });
+      setResult({ error: (err as Error).message });
     } finally {
       setLoading(false);
     }
@@ -32,11 +28,11 @@ const AudiencesPage = () => {
 
   return (
     <>
-      <MetaTags title="Audiences" description="Manage audiences and contacts" />
+      <MetaTags title="Segments" description="Manage segments and contacts" />
 
       <PageHeader
-        title="Audiences"
-        description="Manage audiences and contacts using the Resend API."
+        title="Segments"
+        description="Manage segments and contacts using the Resend API."
       />
 
       <button onClick={fetchContacts} disabled={loading} style={buttonStyle}>
@@ -47,14 +43,17 @@ const AudiencesPage = () => {
 
       <div style={codeBlockStyle}>
         <h3 style={{ margin: "0 0 12px 0", fontSize: "14px" }}>API Code</h3>
-        <pre style={preStyle}>{`// List contacts in an audience
+        <pre style={preStyle}>{`// List segments
+const { data: segments } = await resend.segments.list();
+
+// List contacts in a segment
 const { data, error } = await resend.contacts.list({
-  audienceId: "aud_xxxxxxxxx",
+  segmentId: "seg_xxxxxxxxx",
 });
 
 // Create a contact
 const { data: contact } = await resend.contacts.create({
-  audienceId: "aud_xxxxxxxxx",
+  segmentId: "seg_xxxxxxxxx",
   email: "user@example.com",
   firstName: "John",
 });`}</pre>
@@ -63,4 +62,8 @@ const { data: contact } = await resend.contacts.create({
   );
 };
 
-export default AudiencesPage;
+const buttonStyle: React.CSSProperties = { padding: "10px 20px", backgroundColor: "#18181b", color: "#fff", border: "none", borderRadius: "6px", fontSize: "14px", fontWeight: 500, cursor: "pointer" };
+const codeBlockStyle: React.CSSProperties = { marginTop: "32px", padding: "20px", backgroundColor: "#18181b", borderRadius: "8px", color: "#e4e4e7" };
+const preStyle: React.CSSProperties = { margin: 0, fontSize: "13px", whiteSpace: "pre-wrap", wordBreak: "break-all" };
+
+export default SegmentsPage;
